Revert @2b287f54ae4f9999ab24e85df6ab208dd3f38111
I decided to change the usage of reports_subjects_path back to
report_subject_path because I have found that this method is used
in the specs. I am afraid that this task may not get accepted if I
modify the specs, hence I am resolving the issue of missing
report/subjects route by duplicating routes.