Move checking for tuition from decorator to model
I have noticed that I mistakenly implemented
obviously business logic, and not presentational logic, and that's why
I'm moving it to the student model. I also changed it's implementation.
Now there is a given value (Payment::TUITION) that constitutes what a
"paid" tuition is. I also wrote specs for this method.